fixes #125: BitWriter not emitting all bits
Fixed an issue where BitWriter would drop writes if they aligned on a 64-bit boundary. While
debugging this, I changed uses of arthmetic shift (>>) to logical shift (>>>) which we want to avoid
propagating sign-extended bits.

Took care to ensure a safe left shift is performed when shifting 64-bits. Since java shifts, num %
sizeof(datatype) number of bits, resulting in a no-op when shifting 64-bits which is not what we want.