perf(decoder): read bit sets in larger chunks
Implements `readBits64(int)` with new unit tests, and uses all variants of the
bit reading methods to implement reading into a bit set with larger chunks.

Also adds a `resultShift` parameter to make the implementation of reading vendor
consent much easier, i.e. just need to set this to 1 instead of the default 0.